full/commits_analysis/requirements.txt
kafka-python
requests
python-dateutil
text-generation